Refactored the camera class to split out things that are generic and create CameraBase moved a call to guess_camera_number() inside Camera class. Calls if cam_num=None which is default. Added gphoto2 camera code as DigitalCamera(CameraBase) Interface follows Camera. .get_frame() and .save_frame() outwardly look the same and return same features. .save_frame now returns the filename. The filename by default adds a time and date stamp to filename in both cases. Slight modification of tests.
